Moved time enums to their own folder and namespace.
